Small Fixes
I added a link at the top of the "Search Results" page for a more obvious way of returning to the actual search screen. In addition, I fixed a bug where placing an actor's first name with another actor's last name would break the search query and return nothing. Now, it returns the movies that match with both the mismatching first and last names. This issue has also been fixed for the director search as well.