Finishing Touches on EditDatabase
After making EditDatabase its own page in our app, I failed to realize that anybody can come in and edit the database if they new the URL. Thus, I implemented a countermeasure that checks to make sure that the logged in user is actually a moderator. Also, I commented my code a tiny bit, but not too much since it would become repetitive.