Finalized Search Functionality, Possibly Finished
I made it to where searching for a movie will present more information than just the movie's title. In addition, I have it to where a user can click a link like button to view theater information, as well as in depth rating information. Finally, in that rating popup, a logged in user can leave a rating if they want to (if they are not logged in, then the app will prompt them to do so). With these changes made, I think it's safe to say that the app itself is finished.